Feat: Introduce Product Set Details Page
This commit introduces a new "Product Set Details" page that allows users to view comprehensive information about a Surprise Set. It also enhances the order details screen to correctly display items from these sets.

**Key Changes:**

*   **Feat: Product Set Details Page (`ProductSetDetailsActivity.kt`)**
    *   A new activity, `ProductSetDetailsActivity`, has been created to display details for a Surprise Set.
    *   The layout (`activity_product_set_details.xml`) includes the set name, description, progress bar for sold quantity, and seller information.
    *   It features a tabbed interface to switch between a list of "Products" (available and sold items) and a "How it Works" informational section.
    *   The "Products" tab contains expandable lists for "Items" (available) and "Sold" items, showing quantities and individual item names.

*   **API & ViewModel Integration:**
    *   **`ApiInterface.kt`**: Added a new endpoint `get-set-details` to fetch detailed information for a specific product set.
    *   **`ProductSetDetailsResponse.kt`**: A new response model has been created to handle the data for the set details page.
    *   **`ProductViewModel.kt`**: A new function `getSurpriseProductDetail` is added to call the new API endpoint and expose the data via LiveData.

*   **Enhancement: Order Details for Surprise Sets (`SellerOrderDetailFragment.kt`)**
    *   The seller's order detail screen now correctly displays information for products that are part of a Surprise Set.
    *   If the order is for a set item, it now shows the set name along with the specific item name and its unit ID (e.g., "Item Name #123") instead of a product image.

*   **UI/UX Improvements:**
    *   **Winner Animation (`WatchStreamFragment.kt`)**: Added a popper (confetti) animation to the "Break Spot" winner announcement screen to enhance the user experience.
    *   The "Product Set" card in the live stream is now clickable, navigating the user to the new `ProductSetDetailsActivity`.
    *   **`strings.xml`**: Added a new detailed string resource (`how_it_works_info`) to explain how Surprise Sets work.

+++ app/src/main/java/io/bidswipe/app/network/response/GetOrdersResponse.kt
@@ -263,7 +263,7 @@ data class GetOrdersResponse(
             @SerializedName("payout_id")
             val payoutId: Any?,
             @SerializedName("product_price")
-            val productPrice: Int?,
+            val productPrice: Double?,
             @SerializedName("promote_show_id")
             val promoteShowId: Any?,
             @SerializedName("seller_id")
